Fix play services check
