+++ auth/src/main/java/com/firebase/ui/auth/util/signincontainer/SignInDelegate.java
@@ -1,10 +1,7 @@
 package com.firebase.ui.auth.util.signincontainer;
 
-import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.IntentSender;
-import android.net.ConnectivityManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -15,8 +12,6 @@ import android.text.TextUtils;
 import android.util.Log;
 
 import com.firebase.ui.auth.AuthUI;
-import com.firebase.ui.auth.ErrorCodes;
-import com.firebase.ui.auth.IdpResponse;
 import com.firebase.ui.auth.R;
 import com.firebase.ui.auth.ResultCodes;
 import com.firebase.ui.auth.ui.ExtraConstants;
@@ -27,7 +22,6 @@ import com.firebase.ui.auth.ui.email.RegisterEmailActivity;
 import com.firebase.ui.auth.ui.idp.AuthMethodPickerActivity;
 import com.firebase.ui.auth.util.CredentialsApiHelper;
 import com.firebase.ui.auth.util.GoogleApiConstants;
-import com.firebase.ui.auth.util.PlayServicesHelper;
 import com.google.android.gms.auth.api.Auth;
 import com.google.android.gms.auth.api.credentials.Credential;
 import com.google.android.gms.auth.api.credentials.CredentialRequest;
@@ -64,7 +58,6 @@ public class SignInDelegate extends SmartLockBase<CredentialRequestResult> {
     private static final int RC_IDP_SIGNIN = 3;
     private static final int RC_AUTH_METHOD_PICKER = 4;
     private static final int RC_EMAIL_FLOW = 5;
-    private static final int RC_PLAY_SERVICES = 6;
 
     private Credential mCredential;
 
@@ -73,34 +66,6 @@ public class SignInDelegate extends SmartLockBase<CredentialRequestResult> {
         super.onCreate(savedInstance);
         if (savedInstance != null) {
             // We already have a running instance of this fragment
-            return;
-        }
-
-        if (!hasNetworkConnection()) {
-            Log.d(TAG, "No network connection");
-            finish(ErrorCodes.NO_NETWORK, IdpResponse.getErrorCodeIntent(ErrorCodes.NO_NETWORK));
-            return;
-        }
-
-        // Make Google Play Services available at the correct version, if possible
-        boolean madeAvailable =
-                PlayServicesHelper
-                        .getInstance(getActivity())
-                        .makePlayServicesAvailable(getActivity(), RC_PLAY_SERVICES,
-                                                   new DialogInterface.OnCancelListener() {
-                                                       @Override
-                                                       public void onCancel(DialogInterface dialogInterface) {
-                                                           Log.w(TAG,
-                                                                 "playServices:dialog.onCancel()");
-                                                           finish(ResultCodes.CANCELED,
-                                                                  IdpResponse.getErrorCodeIntent(ErrorCodes.UNKNOWN_ERROR));
-                                                       }
-                                                   });
-
-        if (!madeAvailable
-                || !PlayServicesHelper.getInstance(getActivity()).isPlayServicesAvailable()) {
-            Log.w(TAG, "playServices: could not make available.");
-            finish(ResultCodes.CANCELED, IdpResponse.getErrorCodeIntent(ErrorCodes.UNKNOWN_ERROR));
             return;
         }
 
@@ -184,9 +149,6 @@ public class SignInDelegate extends SmartLockBase<CredentialRequestResult> {
             case RC_AUTH_METHOD_PICKER:
             case RC_EMAIL_FLOW:
                 finish(resultCode, data);
-                break;
-            case RC_PLAY_SERVICES:
-                if (resultCode != ResultCodes.OK) finish(resultCode, data);
                 break;
             default:
                 IdpSignInContainer signInContainer = IdpSignInContainer.getInstance(getActivity());
@@ -353,18 +315,6 @@ public class SignInDelegate extends SmartLockBase<CredentialRequestResult> {
                     RC_IDP_SIGNIN);
             mHelper.dismissDialog();
         }
-    }
-
-    /**
-     * Check if there is an active or soon-to-be-active network connection.
-     */
-    private boolean hasNetworkConnection() {
-        ConnectivityManager manager =
-                (ConnectivityManager) getActivity().getSystemService(Context.CONNECTIVITY_SERVICE);
-
-        return manager != null
-                && manager.getActiveNetworkInfo() != null
-                && manager.getActiveNetworkInfo().isConnectedOrConnecting();
     }
 
     public static void delegate(FragmentActivity activity, FlowParameters params) {
