+++ auth/src/test/java/com/firebase/ui/auth/ui/idp/AuthMethodPickerActivityTest.java
@@ -34,7 +34,6 @@ import com.firebase.ui.auth.testhelpers.LoginManagerShadow;
 import com.firebase.ui.auth.testhelpers.TestConstants;
 import com.firebase.ui.auth.testhelpers.TestHelper;
 import com.firebase.ui.auth.ui.email.RegisterEmailActivity;
-import com.firebase.ui.auth.util.PlayServicesHelper;
 import com.google.firebase.auth.AuthCredential;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FacebookAuthProvider;
@@ -42,7 +41,6 @@ import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.auth.GoogleAuthProvider;
 import com.google.firebase.auth.TwitterAuthProvider;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.Robolectric;
@@ -71,12 +69,6 @@ import static org.mockito.Mockito.when;
                 LoginManagerShadow.class
         }, sdk = 23)
 public class AuthMethodPickerActivityTest {
-
-    @Before
-    public void setUp() {
-        PlayServicesHelper.sApiAvailability = TestHelper.makeMockGoogleApiAvailability();
-    }
-
     @Test
     public void testAllProvidersArePopulated() {
         List<String> providers = Arrays.asList(
